requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dotenv==0.13.0
pytz==2020.1
